Split project into multiple files
