+++ src/search_providers/mod.rs
@@ -1,0 +1,7 @@
+use torrent::Torrent;
+
+pub mod pirate_bay_search;
+
+pub trait SearchProvider {
+    fn search(&self, term: &str) -> Vec<Torrent>;
+}
