+++ src/torrent/mod.rs
@@ -1,0 +1,7 @@
+#[derive(Debug)]
+pub struct Torrent {
+    pub name: String,
+    pub magnet_link: String,
+    pub seeders: Option<u32>,
+    pub leachers: Option<u32>,
+}
